requirements.txt
flask
flask-script
flask-cors
Pillow
pytesseract
gunicorn
gevent
